app/requirements.txt
Flask==2.2
pytesseract==0.3.10
Pillow==8.3.2
Flask-PyMongo==2.3.0
pymongo==4.1.1
celery==5.2.7
redis==4.0.2
validate-docbr==1.7.0
pytest-flask==1.2.0